refactoring microsoftteams js code into logical modules
