+++ src/internal/interfaces.ts
