+++ src/public/publicAPIs.ts
@@ -1,52 +1,11 @@
-import { processMessage, ensureInitialized, sendMessageRequest } from "../internal/MicrosoftTeams.internal";
+import { processMessage, ensureInitialized, sendMessageRequest } from "../internal/internalAPIs";
 import { GlobalVars } from "../internal/GlobalVars";
 import { version, frameContexts } from "../internal/constants";
-import { ExtendedWindow, MessageEvent } from "../internal/MicrosoftTeams.internal.interface";
+import { ExtendedWindow, MessageEvent } from "../internal/interfaces";
 import { settings } from "./settings";
-import { TabInformation, TabInstanceParameters, TabInstance } from "./MicrosoftTeams.public.interface";
+import { TabInformation, TabInstanceParameters, TabInstance, DeepLinkParameters, Context } from "./interfaces";
 
 // ::::::::::::::::::::::: MicrosoftTeams SDK public API ::::::::::::::::::::
-
-export const enum HostClientType {
-  desktop = "desktop",
-  web = "web",
-  android = "android",
-  ios = "ios"
-}
-
-/**
- * Represents information about tabs for an app
- */
-
-
-/**
- * Indicates the team type, currently used to distinguish between different team
- * types in Office 365 for Education (team types 1, 2, 3, and 4).
- */
-export const enum TeamType {
-  Standard = 0,
-  Edu = 1,
-  Class = 2,
-  Plc = 3,
-  Staff = 4
-}
-
-/**
- * Indicates the various types of roles of a user in a team.
- */
-export const enum UserTeamRole {
-  Admin = 0,
-  User = 1,
-  Guest = 2
-}
-
-
-export const enum TaskModuleDimension {
-  Large = "large",
-  Medium = "medium",
-  Small = "small"
-}
-
 /**
  * Initializes the library. This must be called before any other SDK calls
  * but after the frame is loaded successfully.
@@ -357,274 +316,4 @@ export function navigateToTab(tabInstance: TabInstance): void {
       );
     }
   };
-}
-
-export interface Context {
-  /**
-   * The Office 365 group ID for the team with which the content is associated.
-   * This field is available only when the identity permission is requested in the manifest.
-   */
-  groupId?: string;
-
-  /**
-   * The Microsoft Teams ID for the team with which the content is associated.
-   */
-  teamId?: string;
-
-  /**
-   * The name for the team with which the content is associated.
-   */
-  teamName?: string;
-
-  /**
-   * The Microsoft Teams ID for the channel with which the content is associated.
-   */
-  channelId?: string;
-
-  /**
-   * The name for the channel with which the content is associated.
-   */
-  channelName?: string;
-
-  /**
-   * The developer-defined unique ID for the entity this content points to.
-   */
-  entityId: string;
-
-  /**
-   * The developer-defined unique ID for the sub-entity this content points to.
-   * This field should be used to restore to a specific state within an entity, such as scrolling to or activating a specific piece of content.
-   */
-  subEntityId?: string;
-
-  /**
-   * The current locale that the user has configured for the app formatted as
-   * languageId-countryId (for example, en-us).
-   */
-  locale: string;
-
-  /**
-   * @deprecated Use loginHint or userPrincipalName.
-   * The UPN of the current user.
-   * Because a malicious party can run your content in a browser, this value should
-   * be used only as a hint as to who the user is and never as proof of identity.
-   * This field is available only when the identity permission is requested in the manifest.
-   */
-  upn?: string;
-
-  /**
-   * The Azure AD tenant ID of the current user.
-   * Because a malicious party can run your content in a browser, this value should
-   * be used only as a hint as to who the user is and never as proof of identity.
-   * This field is available only when the identity permission is requested in the manifest.
-   */
-  tid?: string;
-
-  /**
-   * The current UI theme.
-   */
-  theme?: string;
-
-  /**
-   * Indication whether the tab is in full-screen mode.
-   */
-  isFullScreen?: boolean;
-
-  /**
-   * The type of the team.
-   */
-  teamType?: TeamType;
-
-  /**
-   * The root SharePoint folder associated with the team.
-   */
-  teamSiteUrl?: string;
-
-  /**
-   * The relative path to the SharePoint folder associated with the channel.
-   */
-  channelRelativeUrl?: string;
-
-  /**
-   * Unique ID for the current Teams session for use in correlating telemetry data.
-   */
-  sessionId?: string;
-
-  /**
-   * The user's role in the team.
-   * Because a malicious party can run your content in a browser, this value should
-   * be used only as a hint as to the user's role, and never as proof of her role.
-   */
-  userTeamRole?: UserTeamRole;
-
-  /**
-   * The Microsoft Teams ID for the chat with which the content is associated.
-   */
-  chatId?: string;
-
-  /**
-   * A value suitable for use as a login_hint when authenticating with Azure AD.
-   * Because a malicious party can run your content in a browser, this value should
-   * be used only as a hint as to who the user is and never as proof of identity.
-   * This field is available only when the identity permission is requested in the manifest.
-   */
-  loginHint?: string;
-
-  /**
-   * The UPN of the current user. This may be an externally-authenticated UPN (e.g., guest users).
-   * Because a malicious party run your content in a browser, this value should
-   * be used only as a hint as to who the user is and never as proof of identity.
-   * This field is available only when the identity permission is requested in the manifest.
-   */
-  userPrincipalName?: string;
-
-  /**
-   * The Azure AD object id of the current user.
-   * Because a malicious party run your content in a browser, this value should
-   * be used only as a hint as to who the user is and never as proof of identity.
-   * This field is available only when the identity permission is requested in the manifest.
-   */
-  userObjectId?: string;
-
-  /**
-   * Indicates whether team is archived.
-   * Apps should use this as a signal to prevent any changes to content associated with archived teams.
-   */
-  isTeamArchived?: boolean;
-
-  /**
-   * The type of the host client. Possible values are : android, ios, web, desktop
-   */
-  hostClientType?: HostClientType;
-
-  /**
-   * SharePoint context
-   */
-  sharepoint?: any;
-
-  /**
-   * The type of license for the current users tenant.
-   */
-  tenantSKU?: string;
-
-  /**
-   * The license type for the current user.
-   */
-  userLicenseType?: string;
-}
-
-export interface DeepLinkParameters {
-  /**
-   * The developer-defined unique ID for the sub-entity to which this deep link points in the current entity.
-   * This field should be used to restore to a specific state within an entity, such as scrolling to or activating a specific piece of content.
-   */
-  subEntityId: string;
-
-  /**
-   * The label for the sub-entity that should be displayed when the deep link is rendered in a client.
-   */
-  subEntityLabel: string;
-
-  /**
-   * The fallback URL to which to navigate the user if the client cannot render the page.
-   * This URL should lead directly to the sub-entity.
-   */
-  subEntityWebUrl?: string;
-}
-
-export interface TaskInfo {
-  /**
-   * The url to be rendered in the webview/iframe.
-   */
-  url?: string;
-
-  /**
-   * JSON defining an adaptive card.
-   */
-  card?: string;
-
-  /**
-   * The requested height of the webview/iframe.
-   */
-  height?: TaskModuleDimension | Number;
-
-  /**
-   * The requested width of the webview/iframe.
-   */
-  width?: TaskModuleDimension | Number;
-
-  /**
-   * Title of the task module.
-   */
-  title?: string;
-
-  /**
-   * If client doesnt support the URL, the URL that needs to be opened in the browser.
-   */
-  fallbackUrl?: string;
-
-  /**
-   * Specifies a bot ID to send the result of the user's interaction with the task module.
-   * If specified, the bot will receive a task/complete invoke event with a JSON object
-   * in the event payload.
-   */
-  completionBotId?: string;
-}
-
-/**
- * Namespace to interact with the task module-specific part of the SDK.
- * This object is usable only on the content frame.
- */
-export namespace tasks {
-  /**
-   * Allows an app to open the task module.
-   * @param taskInfo An object containing the parameters of the task module
-   * @param submitHandler Handler to call when the task module is completed
-   */
-  export function startTask(
-    taskInfo: TaskInfo,
-    submitHandler?: (err: string, result: string) => void
-  ): void {
-    ensureInitialized(frameContexts.content);
-
-    const messageId = sendMessageRequest(GlobalVars.parentWindow, "tasks.startTask", [
-      taskInfo
-    ]);
-    GlobalVars.callbacks[messageId] = submitHandler;
-  }
-
-  /**
-   * Update height/width task info properties.
-   * @param taskInfo An object containing width and height properties
-   */
-  export function updateTask(taskInfo: TaskInfo): void {
-    ensureInitialized(frameContexts.content, frameContexts.task);
-    const { width, height, ...extra } = taskInfo;
-
-    if (!Object.keys(extra).length) {
-      sendMessageRequest(GlobalVars.parentWindow, "tasks.updateTask", [taskInfo]);
-    } else {
-      throw new Error(
-        "updateTask requires a taskInfo argument containing only width and height"
-      );
-    }
-  }
-
-  /**
-   * Submit the task module.
-   * @param result Contains the result to be sent to the bot or the app. Typically a JSON object or a serialized version of it
-   * @param appIds Helps to validate that the call originates from the same appId as the one that invoked the task module
-   */
-  export function submitTask(
-    result?: string | object,
-    appIds?: string | string[]
-  ): void {
-    ensureInitialized(frameContexts.content, frameContexts.task);
-
-    // Send tasks.completeTask instead of tasks.submitTask message for backward compatibility with Mobile clients
-    sendMessageRequest(GlobalVars.parentWindow, "tasks.completeTask", [
-      result,
-      Array.isArray(appIds) ? appIds : [appIds]
-    ]);
-  }
 }
